+++ src/core/bigint.h
@@ -1,0 +1,10 @@
+#ifndef FCOIN_BIGINT_H
+#define FCOIN_BIGINT_H
+
+#include "constants.h"
+
+class bigint {
+
+};
+
+#endif //FCOIN_BIGINT_H