+++ src/core/components/transaction.h
@@ -3,49 +3,35 @@
 
 #include "../constants.h"
 #include "../crypto/sha/sha256.h"
+#include "../interface/serializable.h"
 
-struct transaction_input{
-    str hash;
-    str output_hash;
-    str sign;
-    str txid;
-};
-
-struct transaction_output{
-    str hash;
-    str sender;
-    str receiver;
-    u64 amount;
-    str txid;
-};
-
-class transaction {
+class transaction : public serializable{
 private:
-    str txid;
-    str block;
-    vec<str> inputs;
-    vec<str> outputs;
-    u64 time = 0L;
+    hash512 txid            = { 0 };
+    hash256 block           = { 0 };
+    u64 time                = 0L;
+    vec<hash384> inputs;
+    vec<hash384> outputs;
 public:
-    str input_root,output_root;
-    i32 input_size = 0,output_size = 0;
 
     transaction();
-    transaction(
-            const str& parent,
-            const vec<str>& iData,
-            const vec<str>& oData);
+    transaction(const hash256& _block,const vec<hash384>& _inputs,const vec<hash384>& _outputs);
+    transaction(const vec<hash384>& _inputs,const vec<hash384>& _outputs);
 
-    [[nodiscard]] str getTXID() const;
-    [[nodiscard]] str getBlock() const;
-    [[nodiscard]] vec<str> getInputs() const;
-    [[nodiscard]] vec<str> getOutputs() const;
+    void deserialize(bytebuff &) override;
+    bytebuff serialize() override;
+
+    [[nodiscard]] hash512 getTXID() const;
+    [[nodiscard]] hash256 getBlock() const;
+    [[nodiscard]] vec<hash384> getInputs() const;
+    [[nodiscard]] vec<hash384> getOutputs() const;
     [[nodiscard]] u64 getTime() const;
 
-    void setTXID(const str&);
-    void setBlock(const str&);
-    void setInputs( const vec<str>&);
-    void setOutputs(const vec<str>&);
+private:
+    void setTXID(const hash512 &);
+    void setBlock(const hash256 &);
+    void setInputs( const vec<hash384>&);
+    void setOutputs(const vec<hash384>&);
     void setTime(u64);
 
 };