+++ src/core/wallet.h
@@ -1,13 +1,11 @@
 #ifndef FCOIN_WALLET_H
 #define FCOIN_WALLET_H
 
-#include "./crypto/secp256k1.h"
 #include "components/transaction.h"
 
 class wallet {
 public:
-    explicit wallet(const secp256k1::private_key& privateKey);
-    explicit wallet(const secp256k1::keypair& keys);
+    wallet() = default;
 
     [[nodiscard]] str address() const;
     [[nodiscard]] u64 balance() const;
@@ -17,8 +15,6 @@ public:
     static wallet generate();
 
 private:
-    secp256k1::keypair kp;
-//    utxo_db udb;
 };